+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added 
+
+- Can look up all calendar weeks in current month and current year
+- Can look up all calendar weeks in a certain monthin and current year
+- Can look up all calendar weeks in a certain monthin and certain year
+
 ## [0.1.0] - 2023.06.30
 
 ### Added